Add select
This will allow us to load the data faster when its a larger collection. The downside of this is that you have to list each attribute.

I was trying to use the attributes from swagger.json but those don't always seem to work, I think its because some of those attributes might be virtual